Extract clipboard handling into a hook
